test(logic): split topologyForRegion tests into dedicated file (Refs #2560)
The combined `topology_helpers_test.dart` had grown to 415 physical lines
after the new `topologyForRegion` group, exceeding the
`check_logic_test_file_size` 400-line cap and failing the Quality job on PR #2609.

Extract the new `topologyForRegion (Refs #2560)` group into
`topology_for_region_test.dart` (131 lines). The remaining
`topology_helpers_test.dart` shrinks to 289 lines. No test semantics or
coverage changes.

Verification:
- packages/colonizethis_logic: dart test test/topology_helpers_test.dart
  test/topology_for_region_test.dart -> 14/14 pass.
- dart run tool/check_logic_test_file_size.dart -> no violations.
- dart run tool/check_dart_file_non_comment_line_size.dart -> no violations.
- bash tool/check_test_imports.sh -> passed.
- dart analyze on both files -> clean.

Refs #2560
